# Add BBHash2 marshaling methods