… .vscode/bbhash.txt
… .vscode/settings.json
… bitvector_marshal_test.go
… bitvector_test.go
… internal/test/testname.go