This adds support for the new AppendBinary interface for bitVector, BBHash, and BBHash2 types. The BBHash2 also got MarshalBinary and UnmarshalBinary implementations so that we can actually send it over the network. We also fixed a bug in BBHash.marshalLength that caused trouble for the implementation of BBHash2.

Finally, we added more tests and benchmarks and various code improvements.

Fixes #9

Below we include benchmarks for BBHash's MarshalBinary implementation before and after the introduction of the new AppendBinary method, which makes it easier to write to the buffer. This improves the marshaling
speed significantly.
```
goos: darwin
goarch: arm64
pkg: github.com/relab/bbhash
cpu: Apple M2 Max
                                              │ before-bbhash-marshal.txt │      after-bbhash-marshal.txt       │
                                              │          sec/op           │   sec/op     vs base                │
BBHashMarshalBinary/gamma=1.0/keys=1000-12                    387.3n ± 3%   155.4n ± 1%  -59.88% (p=0.000 n=10)
BBHashMarshalBinary/gamma=1.5/keys=1000-12                    404.3n ± 2%   149.5n ± 1%  -63.02% (p=0.000 n=10)
BBHashMarshalBinary/gamma=2.0/keys=1000-12                    308.3n ± 2%   169.5n ± 1%  -45.02% (p=0.000 n=10)
BBHashMarshalBinary/gamma=1.0/keys=10000-12                  1738.0n ± 1%   845.5n ± 1%  -51.35% (p=0.000 n=10)
BBHashMarshalBinary/gamma=1.5/keys=10000-12                  1423.5n ± 2%   939.4n ± 1%  -34.01% (p=0.000 n=10)
BBHashMarshalBinary/gamma=2.0/keys=10000-12                   1.619µ ± 1%   1.080µ ± 1%  -33.27% (p=0.000 n=10)
BBHashMarshalBinary/gamma=1.0/keys=100000-12                 15.012µ ± 2%   8.534µ ± 1%  -43.15% (p=0.000 n=10)
BBHashMarshalBinary/gamma=1.5/keys=100000-12                 11.352µ ± 3%   8.693µ ± 1%  -23.42% (p=0.000 n=10)
BBHashMarshalBinary/gamma=2.0/keys=100000-12                  13.84µ ± 1%   10.03µ ± 1%  -27.47% (p=0.000 n=10)
BBHashMarshalBinary/gamma=1.0/keys=1000000-12                 65.42µ ± 3%   37.52µ ± 3%  -42.65% (p=0.000 n=10)
BBHashMarshalBinary/gamma=1.5/keys=1000000-12                 55.40µ ± 4%   39.01µ ± 2%  -29.58% (p=0.000 n=10)
BBHashMarshalBinary/gamma=2.0/keys=1000000-12                 62.43µ ± 3%   46.27µ ± 3%  -25.89% (p=0.000 n=10)
geomean                                                       4.654µ        2.729µ       -41.37%

                                              │ before-bbhash-marshal.txt │       after-bbhash-marshal.txt       │
                                              │           B/op            │     B/op      vs base                │
BBHashMarshalBinary/gamma=1.0/keys=1000-12                   1.547Ki ± 0%   1.062Ki ± 0%  -31.31% (p=0.000 n=10)
BBHashMarshalBinary/gamma=1.5/keys=1000-12                   2.039Ki ± 0%   1.062Ki ± 0%  -47.89% (p=0.000 n=10)
BBHashMarshalBinary/gamma=2.0/keys=1000-12                   1.453Ki ± 0%   1.188Ki ± 0%  -18.28% (p=0.000 n=10)
BBHashMarshalBinary/gamma=1.0/keys=10000-12                 12.844Ki ± 0%   8.125Ki ± 0%  -36.74% (p=0.000 n=10)
BBHashMarshalBinary/gamma=1.5/keys=10000-12                 10.039Ki ± 0%   8.750Ki ± 0%  -12.84% (p=0.000 n=10)
BBHashMarshalBinary/gamma=2.0/keys=10000-12                  12.36Ki ± 0%   10.75Ki ± 0%  -13.02% (p=0.000 n=10)
BBHashMarshalBinary/gamma=1.0/keys=100000-12                131.43Ki ± 0%   88.00Ki ± 0%  -33.04% (p=0.000 n=10)
BBHashMarshalBinary/gamma=1.5/keys=100000-12                 95.76Ki ± 0%   88.00Ki ± 0%   -8.10% (p=0.000 n=10)
BBHashMarshalBinary/gamma=2.0/keys=100000-12                 126.0Ki ± 0%   104.0Ki ± 0%  -17.43% (p=0.000 n=10)
BBHashMarshalBinary/gamma=1.0/keys=1000000-12               1247.3Ki ± 0%   752.0Ki ± 0%  -39.71% (p=0.000 n=10)
BBHashMarshalBinary/gamma=1.5/keys=1000000-12                922.4Ki ± 0%   808.0Ki ± 0%  -12.40% (p=0.000 n=10)
BBHashMarshalBinary/gamma=2.0/keys=1000000-12               1154.4Ki ± 0%   912.0Ki ± 0%  -21.00% (p=0.000 n=10)
geomean                                                      39.71Ki        29.63Ki       -25.39%

                                              │ before-bbhash-marshal.txt │      after-bbhash-marshal.txt      │
                                              │         allocs/op         │ allocs/op   vs base                │
BBHashMarshalBinary/gamma=1.0/keys=1000-12                    15.000 ± 0%   2.000 ± 0%  -86.67% (p=0.000 n=10)
BBHashMarshalBinary/gamma=1.5/keys=1000-12                    12.000 ± 0%   2.000 ± 0%  -83.33% (p=0.000 n=10)
BBHashMarshalBinary/gamma=2.0/keys=1000-12                    11.000 ± 0%   2.000 ± 0%  -81.82% (p=0.000 n=10)
BBHashMarshalBinary/gamma=1.0/keys=10000-12                   19.000 ± 0%   2.000 ± 0%  -89.47% (p=0.000 n=10)
BBHashMarshalBinary/gamma=1.5/keys=10000-12                   15.000 ± 0%   2.000 ± 0%  -86.67% (p=0.000 n=10)
BBHashMarshalBinary/gamma=2.0/keys=10000-12                   14.000 ± 0%   2.000 ± 0%  -85.71% (p=0.000 n=10)
BBHashMarshalBinary/gamma=1.0/keys=100000-12                  25.000 ± 0%   2.000 ± 0%  -92.00% (p=0.000 n=10)
BBHashMarshalBinary/gamma=1.5/keys=100000-12                  18.000 ± 0%   2.000 ± 0%  -88.89% (p=0.000 n=10)
BBHashMarshalBinary/gamma=2.0/keys=100000-12                  16.000 ± 0%   2.000 ± 0%  -87.50% (p=0.000 n=10)
BBHashMarshalBinary/gamma=1.0/keys=1000000-12                 29.000 ± 0%   2.000 ± 0%  -93.10% (p=0.000 n=10)
BBHashMarshalBinary/gamma=1.5/keys=1000000-12                 22.000 ± 0%   2.000 ± 0%  -90.91% (p=0.000 n=10)
BBHashMarshalBinary/gamma=2.0/keys=1000000-12                 19.000 ± 0%   2.000 ± 0%  -89.47% (p=0.000 n=10)
geomean                                                        17.24        2.000       -88.40%
```